refactor(ohunt): move some classes
