+++ ohunt/OHunt.Web/Startup.cs
@@ -10,6 +10,7 @@ using Microsoft.AspNet.OData.Extensions;
 using Microsoft.OData.Edm;
 using OHunt.Web.Crawlers;
 using OHunt.Web.Database;
+using OHunt.Web.Dataflow;
 using OHunt.Web.Errors;
 using OHunt.Web.Models;
 using OHunt.Web.Options;
